Prevent rake-task for faking certificates from running in prod

+++ lib/tasks/certificate.rake
@@ -1,6 +1,10 @@
 desc 'Import some random certificate data'
 
 task :generate_certificate do
+  if ENV['STAGE'] == 'production'
+    exit
+  end
+
   ActiveRecord::Base.connection.execute('TRUNCATE TABLE domestic_energy_assessments RESTART IDENTITY')
 
   ActiveRecord::Base.logger = nil
